Extract test trait
